approver: wake phone up fully when we get the call
If phone is in deep sleep mode, the modem wakeup pin will wake phone up,
but in case user has turned screen off and phone is not in deep sleep
mode, wakeup pin will not trigger, and device will ring, but screen will
not wake up. We need to call powerdevil explicitly to wake screen up.