Pause media players on incoming call
When receiving a call you want to be able to talk without any playing music interfering.
Also you want to hear the actual ringtone instead of the music.

Pause all media players when a call is received.
Unpause them when you reject a call.

When the call ends do not unpause the players automatically. This is based on feedback we got for KDE Connect where people were surprised music starts playing on its own after a long phonecall