[contactmapper] Better handle number parsing errors
When the number parsing fails, e.g. because of a syntax error or because
no country has been specified for an non-international number
normalizeNumber returns an empty string.

That empty string will be then used as a key in the map. If multiple
numbers are invalid then contacts are mixed up.

Instead discard invalid numbers before processing. They won't get
matched to a contact, but they will be displayed as-is.

Fixes #28