Properly load call timestamps from database
Currently, the call timestamp appears to be entirely random on the
call list page. I noticed this when looking at #40.

The timestamp is stored into the database as a qint64, so it should be
read out the same way. This changes the way that the number is taken
from the database to eliminate the issue. This same issue has also
occured in Spacebar too:

https://invent.kde.org/plasma-mobile/spacebar/-/commit/6bc8d9229b01925789bb205b425c9f1142ad1ad6