kde-telephony-daemon: refactor MPRIS usage with QDBus
Get rid of qt-mpris since the release of amber-mpris.
Skip amber-mpris integration due to possible option to reuse QCoro for the new MPRIS KDE wrapper if needed.
Move the logic from notification manager to the dialer manager (which also controls audio profiles like speaker and earpiece).

Thanks to original proposal:
https://invent.kde.org/plasma-mobile/plasma-dialer/-/merge_requests/66

Contributes to #24 and:
https://invent.kde.org/teams/plasma-mobile/issues/-/issues/141

Co-Authored-By: Nicolas Fella <nicolas.fella@gmx.de>