[modem-daemon] Pass the USSD state variable through to the user-interface.
Due to the way USSD is currently modelled via DBus, USSD message payload from the network on the one hand and status changes on the other hand is represented by 2 different DBus messages. The user interface needs to support either ordering of these DBus messages. Currently, the user interface did not support the ordering of the payload DBus message first and status DBus message second. This change fixes that by making the user interface aware of the status variable and react to any status variable changes (as well as USSD payload message changes). This has the side effect of possibly rendering the user interface twice (once with an updated USSD payload message without an updated status variable and once again with an updated USSD payload message and with an updated status variable), but the performance impact is negligible given that Kirigami has a high overhead in general.