Fix kwin version check
It looks like KWinEffects was used to check the version of the
installed kwin package. However since Plasma 6 kwin doesn't provide
that cmake package anymore, but it does provide the "KWin" package.

Since it is not used to build but only for a version check, the
package type should be RUNTIME.